BAC-98: Refactoring and clean up code the Auth page
